Separate Gridlines, Axes & Axis from Interaction layer
