+++ src/CollapsedMarks.tsx
@@ -2,10 +2,8 @@ import * as React from 'react'
 import { Marks } from './Marks'
 import { ScaleLinear } from 'd3-scale'
 import { EventComponentFactory, TimelineEvent } from './model'
-import { Axis } from './Axis'
 
 interface Props<EID extends string, LID extends string, E extends TimelineEvent<EID, LID>> {
-  mouseCursor: React.ReactNode
   height: number
   events: ReadonlyArray<E>
   timeScale: ScaleLinear<number, number>
@@ -16,28 +14,29 @@ interface Props<EID extends string, LID extends string, E extends TimelineEvent<
   onEventClick?: (eventId: EID) => void
 }
 
-export const CollapsedMarks = <EID extends string, LID extends string, E extends TimelineEvent<EID, LID>>(
-  props: Props<EID, LID, E>
-) => {
-  const { mouseCursor, height, events, timeScale, eventComponent, onEventHover, onEventUnhover, onEventClick } = props
+export const CollapsedMarks = <EID extends string, LID extends string, E extends TimelineEvent<EID, LID>>({
+  height,
+  events,
+  timeScale,
+  eventComponent,
+  onEventHover,
+  onEventUnhover,
+  onEventClick,
+}: Props<EID, LID, E>) => {
   const y = height / 2
 
   return (
     <g>
-      <Axis y={y} />
-      {mouseCursor}
-      {
-        <Marks
-          height={height}
-          events={events}
-          timeScale={timeScale}
-          y={y}
-          eventComponent={eventComponent}
-          onEventHover={onEventHover}
-          onEventUnhover={onEventUnhover}
-          onEventClick={onEventClick}
-        />
-      }
+      <Marks
+        height={height}
+        events={events}
+        timeScale={timeScale}
+        y={y}
+        eventComponent={eventComponent}
+        onEventHover={onEventHover}
+        onEventUnhover={onEventUnhover}
+        onEventClick={onEventClick}
+      />
     </g>
   )
 }
