+++ src/Timeline.tsx
@@ -15,6 +15,8 @@ import { EventClusters } from './EventClusters'
 import { TimelineTheme } from './theme'
 import { TimelineThemeProvider } from './theme/TimelineThemeProvider'
 import { useZoomLevels } from './hooks/useZoomLevels'
+import { Axes } from './layers/Axes'
+import { Axis } from './layers/Axis'
 
 export interface TimelineProps<EID extends string, LID extends string, E extends TimelineEvent<EID, LID>> {
   width: number
@@ -148,201 +150,197 @@ export const Timeline = <EID extends string, LID extends string, E extends Timel
 
     const showMarks = suppressMarkAnimation ? !isAnimationInProgress : true
 
+    const timeScalePadding = 50
+    const timeScale = scaleLinear()
+      .domain(domain)
+      .range([timeScalePadding, width - timeScalePadding])
+
+    const yScale = scaleBand()
+      .domain(lanes.map((l) => l.laneId))
+      .range([0, height])
+      .paddingInner(0.3)
+      .paddingOuter(0.8)
+
+    const onEventHoverDecorated = (eventId: EID) => {
+      setIsMouseOverEvent(true)
+      onEventHover(eventId)
+    }
+
+    const onEventUnhoverDecorated = (eventId: EID) => {
+      setIsMouseOverEvent(false)
+      onEventUnhover(eventId)
+    }
+
     return (
       <TimelineThemeProvider theme={theme}>
-        <MouseAwareSvg width={width} height={height}>
-          {(mousePosition: SvgCoordinates) => {
-            const timeScalePadding = 50
-            const timeScale = scaleLinear()
-              .domain(domain)
-              .range([timeScalePadding, width - timeScalePadding])
+        <svg viewBox={`0 0 ${width} ${height}`} width={width} height={height}>
+          <GridLines height={height} domain={domain} smallerZoomScale={smallerZoomScale} timeScale={timeScale} />
+          {laneDisplayMode === 'expanded' ? <Axes lanes={lanes} yScale={yScale} /> : <Axis y={height / 2} />}
+          <MouseAwareSvg width={width} height={height}>
+            {(mousePosition: SvgCoordinates) => {
+              const timeAtCursor = timeScale.invert(mousePosition.x)
 
-            const yScale = scaleBand()
-              .domain(lanes.map((l) => l.laneId))
-              .range([0, height])
-              .paddingInner(0.3)
-              .paddingOuter(0.8)
+              const getDomainSpan = (time: number, width: number): Domain => [
+                Math.max(maxDomainStart, time - width / 2),
+                Math.min(maxDomainEnd, time + width / 2),
+              ]
 
-            const timeAtCursor = timeScale.invert(mousePosition.x)
+              const setDomainAnimated = (newDomain: Domain) =>
+                setAnimation({ startMs: Date.now(), fromDomain: domain, toDomain: newDomain })
 
-            const getDomainSpan = (time: number, width: number): Domain => [
-              Math.max(maxDomainStart, time - width / 2),
-              Math.min(maxDomainEnd, time + width / 2),
-            ]
-
-            const setDomainAnimated = (newDomain: Domain) =>
-              setAnimation({ startMs: Date.now(), fromDomain: domain, toDomain: newDomain })
-
-            const updateDomain = (zoomScale: ZoomScale) => () => {
-              if (isDomainChangePossible) {
-                const newZoomWidth = zoomScaleWidth(zoomScale)
-                setDomainAnimated(getDomainSpan(timeAtCursor, newZoomWidth))
-              }
-            }
-
-            const onZoomScrub = () => {
-              if (onCursorMove) {
-                onCursorMove(timeAtCursor, ...getDomainSpan(timeAtCursor, zoomWidth))
-              }
-            }
-
-            const onZoomIn = updateDomain(smallerZoomScale)
-            const onZoomOut = updateDomain(biggerZoomScale)
-
-            const onZoomInCustom = (mouseStartX: number, mouseEndX: number) => {
-              if (isDomainChangePossible) {
-                const newMin = timeScale.invert(mouseStartX)
-                const newMax = timeScale.invert(mouseEndX)
-                setDomainAnimated([newMin, newMax])
-              }
-            }
-
-            const onZoomInCustomInProgress = (mouseStartX: number, mouseEndX: number) => {
-              if (isDomainChangePossible && onCursorMove) {
-                const newMin = timeScale.invert(mouseStartX)
-                const newMax = timeScale.invert(mouseEndX)
-                onCursorMove(newMax, newMin, newMax)
-              }
-            }
-
-            const onZoomReset = () => {
-              if (isDomainChangePossible) {
-                setDomainAnimated([maxDomainStart, maxDomainEnd])
-              }
-            }
-
-            const onPan = (pixelDelta: number) => {
-              if (isDomainChangePossible) {
-                const [domainMin, domainMax] = domain
-                const [rangeMin, rangeMax] = timeScale.range()
-                const domainDelta = (pixelDelta / (rangeMax - rangeMin)) * (domainMax - domainMin)
-                const [newDomainMin, newDomainMax] = [domainMin + domainDelta, domainMax + domainDelta]
-                if (newDomainMin > maxDomainStart && newDomainMax < maxDomainEnd) {
-                  setDomain([newDomainMin, newDomainMax])
+              const updateDomain = (zoomScale: ZoomScale) => () => {
+                if (isDomainChangePossible) {
+                  const newZoomWidth = zoomScaleWidth(zoomScale)
+                  setDomainAnimated(getDomainSpan(timeAtCursor, newZoomWidth))
                 }
               }
-            }
 
-            const onEventHoverDecorated = (eventId: EID) => {
-              setIsMouseOverEvent(true)
-              onEventHover(eventId)
-            }
+              const onZoomScrub = () => {
+                if (onCursorMove) {
+                  onCursorMove(timeAtCursor, ...getDomainSpan(timeAtCursor, zoomWidth))
+                }
+              }
 
-            const onEventUnhoverDecorated = (eventId: EID) => {
-              setIsMouseOverEvent(false)
-              onEventUnhover(eventId)
-            }
+              const onZoomIn = updateDomain(smallerZoomScale)
+              const onZoomOut = updateDomain(biggerZoomScale)
 
-            const [onTrimStart, onTrimEnd] = useTrimming(maxDomain, timeScale, onTrimRangeChange, trimRange)
+              const onZoomInCustom = (mouseStartX: number, mouseEndX: number) => {
+                if (isDomainChangePossible) {
+                  const newMin = timeScale.invert(mouseStartX)
+                  const newMax = timeScale.invert(mouseEndX)
+                  setDomainAnimated([newMin, newMax])
+                }
+              }
 
-            return (
-              <InteractionHandling
-                width={width}
-                height={height}
-                mousePosition={mousePosition}
-                isAnimationInProgress={isAnimationInProgress}
-                isZoomInPossible={isZoomInPossible}
-                isZoomOutPossible={isZoomOutPossible}
-                isTrimming={isTrimming}
-                onHover={onZoomScrub}
-                onZoomIn={onZoomIn}
-                onZoomOut={onZoomOut}
-                onZoomInCustom={onZoomInCustom}
-                onZoomInCustomInProgress={onZoomInCustomInProgress}
-                onZoomReset={onZoomReset}
-                onPan={onPan}
-                onTrimStart={onTrimStart}
-                onTrimEnd={onTrimEnd}
-                onInteractionEnd={onInteractionEnd}
-              >
-                {(cursor, interactionMode, setTrimHoverMode) => {
-                  const mouseCursor =
-                    isNoEventSelected && interactionMode.type !== 'trim' ? (
-                      <MouseCursor
-                        mousePosition={mousePosition.x}
-                        cursorLabel={dateFormat(timeAtCursor)}
-                        cursor={cursor}
-                        interactionMode={interactionMode}
-                        zoomRangeStart={timeScale(timeAtCursor - zoomWidth / 2)!}
-                        zoomRangeEnd={timeScale(timeAtCursor + zoomWidth / 2)!}
-                        zoomScale={smallerZoomScale}
-                        isZoomInPossible={isZoomInPossible}
-                      />
-                    ) : (
-                      <g />
-                    )
+              const onZoomInCustomInProgress = (mouseStartX: number, mouseEndX: number) => {
+                if (isDomainChangePossible && onCursorMove) {
+                  const newMin = timeScale.invert(mouseStartX)
+                  const newMax = timeScale.invert(mouseEndX)
+                  onCursorMove(newMax, newMin, newMax)
+                }
+              }
 
-                  return (
-                    <g>
-                      <GridLines
-                        height={height}
-                        domain={domain}
-                        smallerZoomScale={smallerZoomScale}
-                        timeScale={timeScale}
-                      />
-                      {showMarks && (
-                        <>
-                          <EventClusters
-                            height={height}
-                            eventClusters={eventClustersInsideDomain}
-                            timeScale={timeScale}
-                            yScale={yScale}
-                            expanded={laneDisplayMode === 'expanded'}
+              const onZoomReset = () => {
+                if (isDomainChangePossible) {
+                  setDomainAnimated([maxDomainStart, maxDomainEnd])
+                }
+              }
+
+              const onPan = (pixelDelta: number) => {
+                if (isDomainChangePossible) {
+                  const [domainMin, domainMax] = domain
+                  const [rangeMin, rangeMax] = timeScale.range()
+                  const domainDelta = (pixelDelta / (rangeMax - rangeMin)) * (domainMax - domainMin)
+                  const [newDomainMin, newDomainMax] = [domainMin + domainDelta, domainMax + domainDelta]
+                  if (newDomainMin > maxDomainStart && newDomainMax < maxDomainEnd) {
+                    setDomain([newDomainMin, newDomainMax])
+                  }
+                }
+              }
+
+              const [onTrimStart, onTrimEnd] = useTrimming(maxDomain, timeScale, onTrimRangeChange, trimRange)
+
+              return (
+                <InteractionHandling
+                  width={width}
+                  height={height}
+                  mousePosition={mousePosition}
+                  isAnimationInProgress={isAnimationInProgress}
+                  isZoomInPossible={isZoomInPossible}
+                  isZoomOutPossible={isZoomOutPossible}
+                  isTrimming={isTrimming}
+                  onHover={onZoomScrub}
+                  onZoomIn={onZoomIn}
+                  onZoomOut={onZoomOut}
+                  onZoomInCustom={onZoomInCustom}
+                  onZoomInCustomInProgress={onZoomInCustomInProgress}
+                  onZoomReset={onZoomReset}
+                  onPan={onPan}
+                  onTrimStart={onTrimStart}
+                  onTrimEnd={onTrimEnd}
+                  onInteractionEnd={onInteractionEnd}
+                >
+                  {(cursor, interactionMode, setTrimHoverMode) => {
+                    return (
+                      <g>
+                        {isNoEventSelected && interactionMode.type !== 'trim' ? (
+                          <MouseCursor
+                            mousePosition={mousePosition.x}
+                            cursorLabel={dateFormat(timeAtCursor)}
+                            cursor={cursor}
+                            interactionMode={interactionMode}
+                            zoomRangeStart={timeScale(timeAtCursor - zoomWidth / 2)!}
+                            zoomRangeEnd={timeScale(timeAtCursor + zoomWidth / 2)!}
+                            zoomScale={smallerZoomScale}
+                            isZoomInPossible={isZoomInPossible}
                           />
-                          {laneDisplayMode === 'expanded' ? (
-                            <ExpandedMarks
-                              mouseCursor={mouseCursor}
-                              events={eventsInsideDomain}
-                              lanes={lanes}
-                              timeScale={timeScale}
-                              yScale={yScale}
-                              height={height}
-                              eventComponent={eventComponent}
-                              onEventHover={onEventHoverDecorated}
-                              onEventUnhover={onEventUnhoverDecorated}
-                              onEventClick={onEventClick}
-                            />
-                          ) : (
-                            <CollapsedMarks
-                              mouseCursor={mouseCursor}
-                              events={eventsInsideDomain}
-                              timeScale={timeScale}
-                              height={height}
-                              eventComponent={eventComponent}
-                              onEventHover={onEventHoverDecorated}
-                              onEventUnhover={onEventUnhoverDecorated}
-                              onEventClick={onEventClick}
-                            />
-                          )}
-                        </>
-                      )}
-                      {trimRange && (
-                        <TrimRange
-                          startX={timeScale(trimRange[0])!}
-                          endX={timeScale(trimRange[1])!}
-                          height={height}
-                          width={width}
-                        />
-                      )}
-                      {interactionMode.type === 'trim' && timeScale && (
-                        <Trimmer
-                          startX={trimRange ? trimRange[0] : maxDomain[0]}
-                          endX={trimRange ? trimRange[1] : maxDomain[1]}
-                          height={height}
-                          width={width}
-                          timeScale={timeScale}
-                          setTrimMode={setTrimHoverMode}
-                          dateFormat={dateFormat}
-                          highlightActiveArea={interactionMode.variant === 'trim pan end'}
-                        />
-                      )}
-                    </g>
-                  )
-                }}
-              </InteractionHandling>
-            )
-          }}
-        </MouseAwareSvg>
+                        ) : (
+                          <g />
+                        )}
+                        {trimRange && (
+                          <TrimRange
+                            startX={timeScale(trimRange[0])!}
+                            endX={timeScale(trimRange[1])!}
+                            height={height}
+                            width={width}
+                          />
+                        )}
+                        {interactionMode.type === 'trim' && timeScale && (
+                          <Trimmer
+                            startX={trimRange ? trimRange[0] : maxDomain[0]}
+                            endX={trimRange ? trimRange[1] : maxDomain[1]}
+                            height={height}
+                            width={width}
+                            timeScale={timeScale}
+                            setTrimMode={setTrimHoverMode}
+                            dateFormat={dateFormat}
+                            highlightActiveArea={interactionMode.variant === 'trim pan end'}
+                          />
+                        )}
+                      </g>
+                    )
+                  }}
+                </InteractionHandling>
+              )
+            }}
+          </MouseAwareSvg>
+          <g>
+            {showMarks && (
+              <g>
+                <EventClusters
+                  height={height}
+                  eventClusters={eventClustersInsideDomain}
+                  timeScale={timeScale}
+                  yScale={yScale}
+                  expanded={laneDisplayMode === 'expanded'}
+                />
+                {laneDisplayMode === 'expanded' ? (
+                  <ExpandedMarks
+                    events={eventsInsideDomain}
+                    lanes={lanes}
+                    timeScale={timeScale}
+                    yScale={yScale}
+                    height={height}
+                    eventComponent={eventComponent}
+                    onEventHover={onEventHoverDecorated}
+                    onEventUnhover={onEventUnhoverDecorated}
+                    onEventClick={onEventClick}
+                  />
+                ) : (
+                  <CollapsedMarks
+                    events={eventsInsideDomain}
+                    timeScale={timeScale}
+                    height={height}
+                    eventComponent={eventComponent}
+                    onEventHover={onEventHoverDecorated}
+                    onEventUnhover={onEventUnhoverDecorated}
+                    onEventClick={onEventClick}
+                  />
+                )}
+              </g>
+            )}
+          </g>
+        </svg>
       </TimelineThemeProvider>
     )
   }
